140강 게시물 리스트 페이지 버튼 눌렀을 때 검색조건 안풀리게

+++ src/main/webapp/WEB-INF/jsp/adm/board/list.jsp
@@ -1,6 +1,8 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8"
 	pageEncoding="UTF-8"%>
 
+<%@ page import="com.sbs.untact.util.Util"%>
+
 <%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c"%>
 
 <%@ include file="../part/mainLayoutHead.jspf"%>
@@ -8,7 +10,40 @@
 <section class="section-1">
 	<div class="bg-white shadow-md rounded container mx-auto p-8 mt-8">
 		<div class="flex items-center mb-5">
-		<span class="text-3xl text-black font-bold">게시판 관리</span>
+			<span class="text-3xl text-black font-bold">게시판 관리</span>
+		</div>
+		<hr>
+
+		<!-- 총 게시물 수 -->
+		<div class="flex items-center my-3">
+			<span>총 게시물 수 : ${Util.numberFormat(totleItemsCount)}</span>
+
+			<div class="flex-grow"></div>
+
+			<!-- 검색 -->
+			<form class="flex">
+				<select name="searchKeywordType">
+					<option value="codeAndName">전체</option>
+					<option value="code">코드</option>
+					<option value="name">이름</option>
+				</select>
+
+				<script>
+					if (param.searchKeywordType) {
+						$('.section-1 select[name="searchKeywordType"]').val(
+								param.searchKeywordType);
+					}
+				</script>
+
+				<input
+					class="shadow appearance-none border rounded w-full py-2 px-3 text-grey-darker"
+					autofocus="autofocus" type="text" placeholder="검색어를 입력해주세요"
+					name="searchKeyword" maxlength="20" value="${param.searchKeyword}" />
+				<input
+					class="btn-primary bg-gray-400 text-white font-bold py-2 px-4 rounded"
+					type="submit" value="검색" />
+
+			</form>
 		</div>
 		<hr>
 
@@ -21,24 +56,32 @@
 				<div class="flex justify-between items-center mt-3">
 
 					<!-- 게시판 번호 -->
-					<a href="${detailUrl}" class="font-bold mr-5"><span
-						class="inline-flex justify-center items-center px-3 py-1 rounded-full bg-gray-500 text-white">NO. ${board.id}</span>
-					<span></span></a>
+					<a href="${detailUrl}" class="font-bold mr-5">
+						<span
+							class="inline-flex justify-center items-center px-3 py-1 rounded-full bg-gray-500 text-white">NO.
+							${board.id}</span>
+						<span></span>
+					</a>
 
 					<!-- 등록날짜 -->
 					<span class="font-light text-gray-600">${board.regDate}</span>
-					
+
 					<div class="flex-grow"></div>
+
+					<!-- boardName 공지사항/자유-->
+					<a href="list?boardName=${board.name}"
+						class="px-2 py-1 bg-gray-600 text-gray-100 font-bold rounded hover:bg-gray-500">
+						${board.name} </a>
 				</div>
 
-				<!-- 회원 아아디 -->
+				<!-- 코드 -->
 				<div class="mt-2">
 					<span
 						class="inline-flex justify-center items-center px-3 py-1 rounded-full bg-gray-500 text-white">코드</span>
 					<span>${board.code}</span>
 				</div>
 
-				<!-- 회원 이름 -->
+				<!-- 이름 -->
 				<div class="mt-2">
 					<span
 						class="inline-flex justify-center items-center px-3 py-1 rounded-full bg-gray-500 text-white">이름</span>
@@ -47,15 +90,16 @@
 
 				<div class="flex items-center mt-4 mb-4">
 					<!-- 자세히 보기 -->
-					
-					<a href="${detailUrl}" class="text-gray-700 mr-2 hover:underline"><span>
+
+					<a href="${detailUrl}" class="text-gray-700 mr-2 hover:underline">
+						<span>
 							<i class="fas fa-info"></i>
 							<span class="hidden sm:inline">자세히 보기</span>
-						</span></a>
+						</span>
+					</a>
 
 					<!-- 수정 -->
-					<a href="modify?id=${board.id}"
-						class="text-blue-500 mr-2 hover:underline">
+					<a href="modify?id=${board.id}" class="text-blue-500 mr-2 hover:underline">
 						<span>
 							<i class="fas fa-edit"></i>
 							<span class="hidden sm:inline">수정</span>
@@ -63,13 +107,13 @@
 					</a>
 
 					<!-- 삭제 -->
-					<a onclick="if ( !confirm('삭제하시겠습니까?') ) return false;"
-						href="doDelete?id=${board.id}"
-						class="text-red-500 hover:underline"><span>
+					<a onclick="if ( !confirm('삭제하시겠습니까?') ) return false;" href="doDelete?id=${board.id}" class="text-red-500 hover:underline">
+						<span>
 							<i class="fas fa-trash"></i>
 							<span class="hidden sm:inline">삭제</span>
-						</span></a>
-											
+						</span>
+					</a>
+
 					<div class="flex-grow"></div>
 
 					<!-- 작성자 -->
@@ -85,39 +129,46 @@
 				<hr>
 			</c:forEach>
 		</div>
-		
+
 		<!-- 페이징 -->
-		<nav class="flex justify-center my-5 rounded-md shadow-sm"
-			aria-label="Pagination">
+		<c:set var="pageBtnAddiQueryStr" value="&searchKeywordType=${param.searchKeywordType}&searchKeyword=${param.searchKeyword}" />
+		<nav class="flex justify-center py-5 rounded-md shadow-sm" aria-label="Pagination">
 
 			<!-- 시작 페이지 -->
-			<a href="?page=1"
-				class="relative inline-flex items-center px-2 py-2 rounded-l-md border border-gray-300 bg-white text-sm font-medium text-gray-500 hover:bg-gray-50">
-				<span class="sr-only">Previous</span>
-				<i class="fas fa-chevron-left"></i>
-			</a>
+			<c:if test="${pageMenuStart != 1}">
+				<a href="?page=1" class="relative inline-flex items-center px-2 py-2 rounded-l-md border border-gray-300 bg-white text-sm font-medium text-gray-500 hover:bg-gray-50">
+					<span class="sr-only">Previous</span>
+					<i class="fas fa-chevron-left"></i>
+				</a>
+			</c:if>
 
 			<!-- 페이지 번호 -->
 			<c:forEach var="i" begin="${pageMenuStrat}" end="${pageMenuEnd}">
-				<c:set var="aClassStr"
-					value="relative inline-flex items-center px-4 py-2 border border-gray-300 bg-white text-sm font-medium" />
+				<c:set var="aClassStr" value="relative inline-flex items-center px-4 py-2 border border-gray-300 bg-white text-sm font-medium" />
+
 				<c:set var="aClassStr" value="${aClassStr} active" />
+				
+				<!-- 현재 페이지 -->	
 				<c:if test="${i == page}">
-					<a href="?page=${i}"
-						class="${aClassStr} text-red-700 hover:bg-red-50">${i}</a>
+					<c:set var="aClassStr"
+						value="${aClassStr} text-red-700 hover:bg-red-50" />
 				</c:if>
+
+				<!-- 현재 페이지가 아닌 -->
 				<c:if test="${i != page}">
-					<a href="?page=${i}"
-						class="${aClassStr} text-gray-700 hover:bg-gray-50">${i}</a>
+					<c:set var="aClassStr" value="${aClassStr} text-gray-700 hover:bg-gray-50" />
 				</c:if>
+
+				<a href="?page=${i}${pageBtnAddiQueryStr}" class="${aClassStr}">${i}</a>
 			</c:forEach>
 
 			<!-- 마지막 페이지 -->
-			<a href="?page=${totlePage}"
-				class="relative inline-flex items-center px-2 py-2 rounded-r-md border border-gray-300 bg-white text-sm font-medium text-gray-500 hover:bg-gray-50">
-				<span class="sr-only">Next</span>
-				<i class="fas fa-chevron-right"></i>
-			</a>
+			<c:if test="${pageMenuEnd != totalPage}">
+				<a href="?page=${totlePage}" class="relative inline-flex items-center px-2 py-2 rounded-r-md border border-gray-300 bg-white text-sm font-medium text-gray-500 hover:bg-gray-50">
+					<span class="sr-only">Next</span>
+					<i class="fas fa-chevron-right"></i>
+				</a>
+			</c:if>
 		</nav>
 	</div>
 </section>
